test(api): Replace redundant `moved_app::scope` setup with `moved_app::run` (#359)
