+++ api/src/methods/get_block_by_number.rs
@@ -22,10 +22,10 @@ mod tests {
         super::*,
         crate::methods::tests::create_app,
         alloy::eips::BlockNumberOrTag::{self, *},
-        moved_app::{Command, CommandActor, SpawnWithHandle, TestDependencies},
+        moved_app::{Command, CommandActor, TestDependencies},
         moved_shared::primitives::U64,
         test_case::test_case,
-        tokio::sync::{mpsc, oneshot},
+        tokio::sync::mpsc,
     };
 
     pub fn example_request(tag: BlockNumberOrTag) -> serde_json::Value {
@@ -86,37 +86,29 @@ mod tests {
         let (state_channel, rx) = mpsc::channel(10);
         let (reader, mut app) = create_app();
         let state: CommandActor<TestDependencies> = CommandActor::new(rx, &mut app);
-        let (tx, rx) = oneshot::channel();
 
-        moved_app::scope(|scope| {
-            let state_handle = scope.spawn_with_handle(state.work());
+        moved_app::run(state, async move {
+            let request = example_request(Latest);
+            let response = execute(request, reader.clone()).await.unwrap();
+            assert_eq!(get_block_number_from_response(response), "0x0");
 
-            scope.spawn_with_handle(async move {
-                let request = example_request(Latest);
-                let response = execute(request, reader.clone()).await.unwrap();
-                assert_eq!(get_block_number_from_response(response), "0x0");
+            // Create a block, so the block height becomes 1
+            let msg = Command::StartBlockBuild {
+                payload_attributes: Default::default(),
+                payload_id: U64::from(0x03421ee50df45cacu64),
+            };
+            state_channel.send(msg).await.unwrap();
 
-                // Create a block, so the block height becomes 1
-                let msg = Command::StartBlockBuild {
-                    payload_attributes: Default::default(),
-                    payload_id: U64::from(0x03421ee50df45cacu64),
-                };
-                state_channel.send(msg).await.unwrap();
+            state_channel.reserve_many(10).await.unwrap();
 
-                state_channel.reserve_many(10).await.unwrap();
+            let request = example_request(Latest);
+            let response = execute(request, reader.clone()).await.unwrap();
 
-                let request = example_request(Latest);
-                let response = execute(request, reader.clone()).await.unwrap();
+            assert_eq!(get_block_number_from_response(response), "0x1");
 
-                assert_eq!(get_block_number_from_response(response), "0x1");
-
-                drop(state_channel);
-                state_handle.await.unwrap();
-                tx.send(()).unwrap();
-            });
-        });
-
-        rx.await.unwrap();
+            drop(state_channel);
+        })
+        .await;
     }
 
     #[test_case(Safe; "safe")]
@@ -127,30 +119,20 @@ mod tests {
         let (state_channel, rx) = mpsc::channel(10);
         let (reader, mut app) = create_app();
         let state: CommandActor<TestDependencies> = CommandActor::new(rx, &mut app);
-        let (tx, rx) = oneshot::channel();
 
-        moved_app::scope(|scope| {
-            let state_handle = scope.spawn_with_handle(state.work());
+        moved_app::run(state, async move {
+            let msg = Command::StartBlockBuild {
+                payload_attributes: Default::default(),
+                payload_id: U64::from(0x03421ee50df45cacu64),
+            };
+            state_channel.send(msg).await.unwrap();
 
-            scope.spawn_with_handle(async move {
-                let msg = Command::StartBlockBuild {
-                    payload_attributes: Default::default(),
-                    payload_id: U64::from(0x03421ee50df45cacu64),
-                };
-                state_channel.send(msg).await.unwrap();
+            state_channel.reserve_many(10).await.unwrap();
 
-                state_channel.reserve_many(10).await.unwrap();
-
-                let request = example_request(tag);
-                let response = execute(request, reader.clone()).await.unwrap();
-                assert_eq!(get_block_number_from_response(response), "0x1");
-
-                drop(state_channel);
-                state_handle.await.unwrap();
-                tx.send(()).unwrap();
-            });
-        });
-
-        rx.await.unwrap();
+            let request = example_request(tag);
+            let response = execute(request, reader.clone()).await.unwrap();
+            assert_eq!(get_block_number_from_response(response), "0x1");
+        })
+        .await;
     }
 }
