+++ api/src/methods/send_raw_transaction.rs
@@ -51,10 +51,7 @@ async fn inner_execute(tx: TxEnvelope, queue: CommandQueue) -> Result<B256, Json
 
 #[cfg(test)]
 pub mod tests {
-    use {
-        super::*, crate::methods::tests::create_app, moved_app::SpawnWithHandle,
-        tokio::sync::oneshot,
-    };
+    use {super::*, crate::methods::tests::create_app};
 
     pub fn example_request() -> serde_json::Value {
         serde_json::from_str(
@@ -75,27 +72,19 @@ pub mod tests {
     async fn test_execute() {
         let (_reader, mut app) = create_app();
         let (queue, state) = moved_app::create(&mut app, 10);
-        let (tx, rx) = oneshot::channel();
 
-        moved_app::scope(|scope| {
-            let state_handle = scope.spawn_with_handle(state.work());
+        moved_app::run(state, async move {
+            let request = example_request();
 
-            scope.spawn_with_handle(async move {
-                let request = example_request();
+            let expected_response: serde_json::Value = serde_json::from_str(
+                r#""0x3545efb3ce7a22353c346c98771640131b81baa64eb03113b20ad2bef5c0ec53""#,
+            )
+            .unwrap();
 
-                let expected_response: serde_json::Value = serde_json::from_str(
-                    r#""0x3545efb3ce7a22353c346c98771640131b81baa64eb03113b20ad2bef5c0ec53""#,
-                )
-                .unwrap();
+            let response = execute(request, queue).await.unwrap();
 
-                let response = execute(request, queue).await.unwrap();
-
-                assert_eq!(response, expected_response);
-                state_handle.await.unwrap();
-                tx.send(()).unwrap();
-            });
-        });
-
-        rx.await.unwrap();
+            assert_eq!(response, expected_response);
+        })
+        .await;
     }
 }
